requirements.txt
Flask
pymongo
Flask-PyMongo
pyjwt
gunicorn
pytest
pylint